Add test_2d_gpu_variable
