+++ oneflow/python/framework/c_api_util.py
@@ -1,5 +1,7 @@
 from __future__ import absolute_import
 
+from google.protobuf import text_format
+
 import oneflow.core.common.data_type_pb2 as dtype_util
 import oneflow.core.common.error_pb2 as error_util
 import oneflow.core.job.env_pb2 as env_pb2
@@ -11,7 +13,6 @@ import oneflow.core.operator.op_conf_pb2 as op_conf_util
 import oneflow.core.record.record_pb2 as record_util
 import oneflow.core.register.logical_blob_id_pb2 as logical_blob_id_util
 import oneflow.oneflow_internal as oneflow_internal
-from google.protobuf import text_format
 from oneflow.core.framework.config_def_pb2 import ConfigDef
 from oneflow.core.job.inter_user_job_info_pb2 import InterUserJobInfo
 from oneflow.python.framework.job_build_and_infer_error import JobBuildAndInferError
@@ -19,13 +20,6 @@ from oneflow.python.framework.job_build_and_infer_error import JobBuildAndInferE
 
 def RegisterWatcherOnlyOnce(watcher):
     error_str = oneflow_internal.RegisterWatcherOnlyOnce(watcher)
-    error = text_format.Parse(error_str, error_util.ErrorProto())
-    if error.HasField("error_type"):
-        raise JobBuildAndInferError(error)
-
-
-def RegisterWorkerCallbackOnlyOnce(callback):
-    error_str = oneflow_internal.RegisterWorkerCallbackOnlyOnce(callback)
     error = text_format.Parse(error_str, error_util.ErrorProto())
     if error.HasField("error_type"):
         raise JobBuildAndInferError(error)
